projects/user-score/requirements.txt
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
joblib==1.3.0
pyyaml==6.0.0
fsspec==2024.6.0
s3fs==2024.6.0
